Split out to rule_types.py
